Improve vc gui and add icons for more things

+++ SerbleChat.Backend/Services/IVoiceManager.cs
@@ -3,6 +3,6 @@ using Livekit.Server.Sdk.Dotnet;
 namespace SerbleChat.Backend.Services;
 
 public interface IVoiceManager {
-    public Task<List<string>> GetConnectedUsers(int channelId);
+    public Task<List<string>> GetConnectedUsers(long channelId);
     public Task OnWebhook(WebhookEvent ev);
 }
